requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
drf-yasg
python-dotenv
gunicorn